telegram_bot_v2/requirements.txt
# Core Deep Learning dependencies
torch==2.0.0
torchvision==0.15.1

# Real-ESRGAN dependencies
basicsr==1.4.2
facexlib==0.2.5
gfpgan==1.3.8
realesrgan==0.3.0

# Image processing
opencv-python==4.6.0.66
numpy==1.23.5
Pillow==9.2.0

# Utility and other requirements
scipy==1.9.3
tqdm==4.64.1
lmdb==1.3.0
pyyaml==6.0

# Optional but sometimes needed
requests==2.28.1
scikit-image==0.19.3


python-telegram-bot==20.6
huggingface-hub==0.27.0
ultralytics==8.3.55
ultralytics-thop==2.0.13
pillow_heif==0.15.0
openai==1.58.1
python-dotenv==1.0.1
aiohttp==3.10.11